+++ scripts/helpers/ui/spellbook-filters.mjs
@@ -2,7 +2,7 @@ import { log } from '../../logger.mjs';
 import * as filterUtils from '../filters.mjs';
 
 /**
- * Helper class for filtering spells in the spellbook application
+ * Helper class for filtering spells in the spellbook application with cached filter state
  */
 export class SpellbookFilterHelper {
   /**
@@ -12,6 +12,8 @@ export class SpellbookFilterHelper {
   constructor(app) {
     this.app = app;
     this.actor = app.actor;
+    this._cachedFilterState = null;
+    this._lastFilterUpdate = 0;
   }
 
   /**
@@ -23,13 +25,23 @@ export class SpellbookFilterHelper {
   }
 
   /**
-   * Get the current filter state from the UI
+   * Invalidate cached filter state
+   */
+  invalidateFilterCache() {
+    this._cachedFilterState = null;
+    this._lastFilterUpdate = 0;
+  }
+
+  /**
+   * Get the current filter state from the UI (with caching)
    * @returns {Object} The filter state
    */
   getFilterState() {
+    const now = Date.now();
+    if (this._cachedFilterState && now - this._lastFilterUpdate < 100) return this._cachedFilterState;
     if (!this.element) return filterUtils.getDefaultFilterState();
 
-    return {
+    this._cachedFilterState = {
       name: this.element.querySelector('[name="filter-name"]')?.value || '',
       level: this.element.querySelector('[name="filter-level"]')?.value || '',
       school: this.element.querySelector('[name="filter-school"]')?.value || '',
@@ -45,6 +57,194 @@ export class SpellbookFilterHelper {
       materialComponents: this.element.querySelector('[name="filter-materialComponents"]')?.value || '',
       sortBy: this.element.querySelector('[name="sort-by"]')?.value || 'level'
     };
+
+    this._lastFilterUpdate = now;
+    return this._cachedFilterState;
+  }
+
+  /**
+   * Filter available spells based on current filter state (moved from GMSpellListManager)
+   * @param {Array} availableSpells - Array of available spells
+   * @param {Set} selectedSpellUUIDs - Set of selected spell UUIDs
+   * @param {Function} isSpellInSelectedList - Function to check if spell is in selected list
+   * @returns {Object} Filtered spells with count
+   */
+  filterAvailableSpells(availableSpells, selectedSpellUUIDs, isSpellInSelectedList) {
+    try {
+      const filterState = this.getFilterState();
+      log(3, 'Beginning Filtering:', selectedSpellUUIDs.size, 'selected spells out of', availableSpells.length, 'total available');
+      let remainingSpells = [...availableSpells];
+      remainingSpells = this._filterBySelectedList(remainingSpells, selectedSpellUUIDs, isSpellInSelectedList);
+      remainingSpells = this._filterBySource(remainingSpells, filterState);
+      remainingSpells = this._filterByBasicProperties(remainingSpells, filterState);
+      remainingSpells = this._filterByRange(remainingSpells, filterState);
+      remainingSpells = this._filterByDamageAndConditions(remainingSpells, filterState);
+      remainingSpells = this._filterBySpecialProperties(remainingSpells, filterState);
+      log(3, 'Final spells count:', remainingSpells.length);
+      return { spells: remainingSpells, totalFiltered: remainingSpells.length };
+    } catch (error) {
+      log(1, 'Error filtering available spells:', error);
+      return { spells: [], totalFiltered: 0 };
+    }
+  }
+
+  /**
+   * Filter out spells already in the selected list
+   * @param {Array} spells - Spells to filter
+   * @param {Set} selectedSpellUUIDs - UUIDs in selected list
+   * @param {Function} isSpellInSelectedList - Function to check if spell is in list
+   * @returns {Array} Filtered spells
+   * @private
+   */
+  _filterBySelectedList(spells, selectedSpellUUIDs, isSpellInSelectedList) {
+    const filtered = spells.filter((spell) => !isSpellInSelectedList(spell, selectedSpellUUIDs));
+    log(3, 'After in-list filter:', filtered.length, 'spells remaining');
+    return filtered;
+  }
+
+  /**
+   * Filter spells by source
+   * @param {Array} spells - Spells to filter
+   * @param {Object} filterState - Current filter state
+   * @returns {Array} Filtered spells
+   * @private
+   */
+  _filterBySource(spells, filterState) {
+    const { source } = filterState;
+    if (!source || source.trim() === '' || source === 'all') return spells;
+    const beforeCount = spells.length;
+    const filtered = spells.filter((spell) => {
+      const spellSource = (spell.sourceId || '').split('.')[0];
+      const packName = spell.packName || '';
+      return spellSource.includes(source) || spellSource === source || packName.toLowerCase().includes(source.toLowerCase());
+    });
+
+    if (filtered.length === 0 && beforeCount > 0) {
+      log(3, `Source '${source}' filtered out all spells, resetting to show all sources`);
+      filterState.source = 'all';
+      return spells;
+    }
+
+    log(3, `After source filter: ${filtered.length} spells remaining`);
+    return filtered;
+  }
+
+  /**
+   * Filter spells by basic properties (name, level, school, casting time)
+   * @param {Array} spells - Spells to filter
+   * @param {Object} filterState - Current filter state
+   * @returns {Array} Filtered spells
+   * @private
+   */
+  _filterByBasicProperties(spells, filterState) {
+    const { name, level, school, castingTime } = filterState;
+    let filtered = spells;
+    if (name) filtered = filtered.filter((spell) => spell.name.toLowerCase().includes(name.toLowerCase()));
+    if (level) {
+      const levelValue = parseInt(level);
+      filtered = filtered.filter((spell) => spell.level === levelValue);
+    }
+    if (school) filtered = filtered.filter((spell) => spell.school === school);
+    if (castingTime) {
+      filtered = filtered.filter((spell) => {
+        const [filterType, filterValue] = castingTime.split(':');
+        const spellCastingType = spell.filterData?.castingTime?.type || spell.system?.activation?.type || '';
+        const spellCastingValue = String(spell.filterData?.castingTime?.value || spell.system?.activation?.value || '1');
+        return spellCastingType === filterType && spellCastingValue === filterValue;
+      });
+    }
+
+    return filtered;
+  }
+
+  /**
+   * Filter spells by range
+   * @param {Array} spells - Spells to filter
+   * @param {Object} filterState - Current filter state
+   * @returns {Array} Filtered spells
+   * @private
+   */
+  _filterByRange(spells, filterState) {
+    const { minRange, maxRange } = filterState;
+    if (!minRange && !maxRange) return spells;
+    const filtered = spells.filter((spell) => {
+      if (!(spell.filterData?.range?.units || spell.system?.range?.units)) return true;
+      const rangeUnits = spell.filterData?.range?.units || spell.system?.range?.units || '';
+      const rangeValue = parseInt(spell.system?.range?.value || 0);
+      let standardizedRange = rangeValue;
+      if (rangeUnits === 'mi') standardizedRange = rangeValue * 5280;
+      else if (rangeUnits === 'spec') standardizedRange = 0;
+      const minRangeVal = minRange ? parseInt(minRange) : 0;
+      const maxRangeVal = maxRange ? parseInt(maxRange) : Infinity;
+      return standardizedRange >= minRangeVal && standardizedRange <= maxRangeVal;
+    });
+
+    log(3, 'After range filter:', filtered.length, 'spells remaining');
+    return filtered;
+  }
+
+  /**
+   * Filter spells by damage types and conditions
+   * @param {Array} spells - Spells to filter
+   * @param {Object} filterState - Current filter state
+   * @returns {Array} Filtered spells
+   * @private
+   */
+  _filterByDamageAndConditions(spells, filterState) {
+    const { damageType, condition } = filterState;
+    let filtered = spells;
+
+    if (damageType) {
+      filtered = filtered.filter((spell) => {
+        const spellDamageTypes = Array.isArray(spell.filterData?.damageTypes) ? spell.filterData.damageTypes : [];
+        return spellDamageTypes.length > 0 && spellDamageTypes.includes(damageType);
+      });
+    }
+
+    if (condition) {
+      filtered = filtered.filter((spell) => {
+        const spellConditions = Array.isArray(spell.filterData?.conditions) ? spell.filterData.conditions : [];
+        return spellConditions.includes(condition);
+      });
+    }
+
+    return filtered;
+  }
+
+  /**
+   * Filter spells by special properties (saves, concentration, ritual)
+   * @param {Array} spells - Spells to filter
+   * @param {Object} filterState - Current filter state
+   * @returns {Array} Filtered spells
+   * @private
+   */
+  _filterBySpecialProperties(spells, filterState) {
+    const { requiresSave, concentration, ritual, materialComponents } = filterState;
+    let filtered = spells;
+
+    if (requiresSave) {
+      filtered = filtered.filter((spell) => {
+        const spellRequiresSave = spell.filterData?.requiresSave || false;
+        return (requiresSave === 'true' && spellRequiresSave) || (requiresSave === 'false' && !spellRequiresSave);
+      });
+    }
+
+    if (concentration) {
+      filtered = filtered.filter((spell) => {
+        const requiresConcentration = !!spell.filterData?.concentration;
+        return (concentration === 'true' && requiresConcentration) || (concentration === 'false' && !requiresConcentration);
+      });
+    }
+
+    if (materialComponents) {
+      filtered = filtered.filter((spell) => {
+        const hasMaterialComponents = spell.filterData?.materialComponents?.hasConsumedMaterials || false;
+        return (materialComponents === 'consumed' && hasMaterialComponents) || (materialComponents === 'notConsumed' && !hasMaterialComponents);
+      });
+    }
+
+    if (ritual) filtered = filtered.filter((spell) => !!spell.filterData?.isRitual);
+    return filtered;
   }
 
   /**
@@ -244,7 +444,7 @@ export class SpellbookFilterHelper {
           countablePrepared: 0
         };
 
-        container.style.display = levelStats.visible > 0 ? '' : ''; // TODO: Fix this when '' = none, our new templates changed how this is found
+        container.style.display = levelStats.visible > 0 ? '' : '';
         const countDisplay = container.querySelector('.spell-count');
         if (countDisplay && levelStats.countable > 0) {
           countDisplay.textContent = `(${levelStats.countablePrepared}/${levelStats.countable})`;
